refactor: complete evaluator extensibility
- Moves result processing to each evaluator
- Add BaseEvaluator interface with process_results/show_results
- Easy addition of new evaluator types
